+++ dart_native/android/src/main/jni/dn_type_convert.cpp
@@ -6,79 +6,86 @@
 #include "dn_type_convert.h"
 #include "dn_log.h"
 
-jstring convertToJavaUtf16(JNIEnv *env, void *value, jvalue *argValue, int index)
-{
-  uint16_t *utf16 = (uint16_t *)value;
+#define SET_JAVA_VALUE(jtype, ctype, indexType) \
+  bool Set##jtype##Value(void *value, jvalue *argValue, int index) { \
+    argValue[index].indexType = (jtype) * (ctype *)value; \
+    return sizeof(void *) == 4 && sizeof(ctype) > 4; \
+  }
+
+SET_JAVA_VALUE(jint, int, i)
+SET_JAVA_VALUE(jchar, char, c)
+SET_JAVA_VALUE(jbyte, int8_t, b)
+SET_JAVA_VALUE(jshort, int16_t, s)
+SET_JAVA_VALUE(jlong, long long, j)
+SET_JAVA_VALUE(jfloat, float, f)
+SET_JAVA_VALUE(jdouble, double, d)
+
+bool convertToJBoolean(void *value, jvalue *argValue, int index) {
+  argValue[index].z = static_cast<jboolean>(*((int *) value));
+  return false;
+}
+
+/// key is basic argument signature
+static const std::map<char, std::function<BasicTypeToNative>>
+    basicTypeConvertMap = {
+    {'C', SetjcharValue},
+    {'I', SetjintValue},
+    {'D', SetjdoubleValue},
+    {'F', SetjfloatValue},
+    {'B', SetjbyteValue},
+    {'S', SetjshortValue},
+    {'J', SetjlongValue},
+    {'Z', convertToJBoolean}};
+
+
+std::map<char, std::function<BasicTypeToNative>> GetTypeConvertMap() {
+  return basicTypeConvertMap;
+}
+
+jstring convertToJavaUtf16(JNIEnv *env, void *value) {
+  auto *utf16 = (uint16_t *) value;
 
   uint32_t length = 0;
   length += *utf16++ << 16;
   length += *utf16++;
 
-  jstring nativeString = env->NewString(utf16, length);
+  auto nativeString = env->NewString(utf16, length);
   free(value);
 
-  if (argValue != nullptr)
-  {
-    argValue[index].l = nativeString;
-  }
   return nativeString;
 }
 
-uint16_t *convertToDartUtf16(JNIEnv *env, jstring nativeString)
-{
-  const jchar *jc = env->GetStringChars(nativeString, NULL);
+/// nativeString not null
+uint16_t *convertToDartUtf16(JNIEnv *env, jstring nativeString) {
+  const jchar *jc = env->GetStringChars(nativeString, nullptr);
   jsize strLength = env->GetStringLength(nativeString);
 
-  uint16_t *utf16Str = new uint16_t[strLength + 3];
+  /// check bom
+  bool hasBom = jc[0] == 0xFEFF || jc[0] == 0xFFFE; // skip bom
+  int indexStart = 0;
+  if (hasBom) {
+    strLength--;
+    indexStart = 1;
+    if (strLength <= 0) {
+      env->ReleaseStringChars(nativeString, jc);
+      env->DeleteLocalRef(nativeString);
+      return nullptr;
+    }
+  }
+
+  /// do convert
+  auto *utf16Str =
+      static_cast<uint16_t *>(malloc(sizeof(uint16_t) * (strLength + 3)));
+  /// save u16list length
   utf16Str[0] = strLength >> 16 & 0xFFFF;
   utf16Str[1] = strLength & 0xFFFF;
-
-  for (int i = 0; i < strLength; i++)
-  {
-    utf16Str[i + 2] = jc[i];
+  int u16Index = 2;
+  for (int i = indexStart; i < strLength; i++) {
+    utf16Str[u16Index++] = jc[i];
   }
   utf16Str[strLength + 2] = '\0';
 
   env->ReleaseStringChars(nativeString, jc);
+  env->DeleteLocalRef(nativeString);
   return utf16Str;
-}
-
-void convertToJChar(void *value, jvalue *argValue, int index)
-{
-  argValue[index].c = (jchar) * (char *)value;
-}
-
-void convertToJInt(void *value, jvalue *argValue, int index)
-{
-  argValue[index].i = (jint) * (int *)value;
-}
-
-void convertToJDouble(void *value, jvalue *argValue, int index)
-{
-  argValue[index].d = (jdouble) * (double *)value;
-}
-
-void convertToJFloat(void *value, jvalue *argValue, int index)
-{
-  argValue[index].f = (jfloat) * (float *)value;
-}
-
-void convertToJByte(void *value, jvalue *argValue, int index)
-{
-  argValue[index].b = (jbyte) * (int8_t *)value;
-}
-
-void convertToJShort(void *value, jvalue *argValue, int index)
-{
-  argValue[index].s = (jshort) * (int16_t *)value;
-}
-
-void convertToJLong(void *value, jvalue *argValue, int index)
-{
-  argValue[index].j = (jlong) * (long *)value;
-}
-
-void convertToJBoolean(void *value, jvalue *argValue, int index)
-{
-  argValue[index].z = static_cast<jboolean>(*((int *)value));
 }