+++ dart_native/example/lib/android/unit_test.dart
@@ -1,11 +1,28 @@
+import 'package:dart_native_example/android/delegate_stub.dart';
 import 'package:dart_native_example/android/runtimestub.dart';
 import 'package:dart_native_example/android/entity.dart';
 
 import 'delegate_stub.dart';
+import 'package:dart_native_example/dn_unit_test.dart';
+
+/// Android unit test implementation.
+class DNAndroidUnitTest with DNUnitTestBase {
+  final stub = RuntimeStub();
+
+  @override
+  String fooString(String str) {
+    return stub.getString(str);
+  }
+
+  @override
+  void runAllUnitTests() {
+    testAndroid(stub);
+  }
+}
 
 testAndroid(RuntimeStub stub) {
   int ms = currentTimeMillis();
-  double resultDouble = stub.getDouble(10.0);
+  double resultDouble = stub.getDouble(3.40282e+038);
   int use = currentTimeMillis() - ms;
   print('getDouble result:$resultDouble , cost:$use');
 
@@ -40,7 +57,7 @@ testAndroid(RuntimeStub stub) {
   print('getShort result:$resultShort , cost:$use');
 
   ms = currentTimeMillis();
-  int resultLong = stub.getLong(100);
+  int resultLong = stub.getLong(4294967296);
   use = currentTimeMillis() - ms;
   print('getLong result:$resultLong , cost:$use');
 
@@ -121,7 +138,6 @@ testAndroid(RuntimeStub stub) {
   }
 
   stub.setDelegateListener(DelegateStub());
-
 }
 
 int currentTimeMillis() {