+++ dart_native/example/lib/ios/unit_test.dart
@@ -3,8 +3,26 @@ import 'dart:ffi';
 import 'dart:isolate';
 
 import 'package:dart_native/dart_native.dart';
+import 'package:dart_native_example/dn_unit_test.dart';
 import 'package:dart_native_example/ios/delegatestub.dart';
+import 'package:dart_native_example/ios/runtimeson.dart';
 import 'package:dart_native_example/ios/runtimestub.dart';
+
+/// IOS unit test implementation.
+class DNIOSUnitTest with DNUnitTestBase {
+  final stub = RuntimeSon();
+  final delegate = DelegateStub();
+
+  @override
+  String fooString(String str) {
+    return stub.fooNSString(str);
+  }
+
+  @override
+  void runAllUnitTests() {
+    testIOS(stub, delegate);
+  }
+}
 
 testIOS(RuntimeStub stub, DelegateStub delegate) {
   bool resultBool = stub.fooBOOL(false);